pyproject.toml
[project]
name = "urbanemissions-rag"
version = "0.1.0"
description = "RAG chat interface for urbanemissions.info"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "beautifulsoup4",
    "lxml",
    "chromadb",
    "sentence-transformers",
    "google-genai",
    "python-dotenv",
    "pydantic",
]
